requirements.prod.txt
django
djangorestframework
jsonfield
user-agents
